+++ CHANGELOG.md
@@ -1,8 +1,22 @@
 # Changelog
 
+## [0.2.0]
+
+- Added optional `-r` (ROLE) and `-w` (WAREHOUSE) arguments for SnowDDL CLI interface.
+- Added basic converters from existing `DATABASE`, `SCHEMA`, `TABLE`, `SEQUENCE` objects to SnowDDL YAML configs.
+- Removed future grants from `SCHEMA ROLE (WRITE)` for `VIEW` object type.
+- "Getting Started Test" workflow now runs each config version twice to detect possible changes being missed on first run.
+- Fixed bug with `comment` and `default` not being applied to `TABLE` columns in some cases.
+- Fixed bug with short hashes being used as byte-strings instead of properly decoded pure `ascii` representations.
+- Fixed bug with other alters being applied to columns dropped from `TABLE` in some cases.
+- Fixed bug with table column comment not being applied immediately on `ADD COLUMN`.
+- Fixed bug with role comment not being applied immediately on `CREATE ROLE`.
+- Reworked the way how `comment` is being applied to `VIEW` object type. Snowflake implicitly modifies view `text` in `SHOW VIEWS` if it contains a `comment` during `CREATE VIEW`, which breaks view checks on subsequent runs.
+- If `VIEW` was replaced, the resolve result is now `REPLACE` instead of `ALTER`.
+
 ## [0.1.1] - 2022-02-11
 
-- Fix typing annotations for `List` and `Dict` to make it compatible with Python 3.7
+- Fixed typing annotations for `List` and `Dict` to make it compatible with Python 3.7.
 
 ## [0.1.0] - 2022-02-10